Tidied up the code
Updated the code to make it a bit more reliable.
Updated YouTube Extractor
Added NAudio

+++ iPhoneRingToneMaker/RingtoneMakerIPhone.cs
@@ -1,47 +1,45 @@
 ﻿using System;
-using System.Collections.Generic;
 using System.ComponentModel;
-using System.Data;
 using System.Diagnostics;
-using System.Drawing;
 using System.IO;
 using System.Linq;
-using System.Text;
-using System.Web;
 using System.Windows.Forms;
-
+using NAudio.Wave;
 using iTunesLib;
 using YoutubeExtractor;
 
 namespace iPhoneRingtoneMaker
 {
-    public partial class iPhoneRingtoneMaker : Form
+    public partial class RingtoneMakerIPhone : Form
     {
-        public static string folder = null;
-        public iTunesApp itunes = new iTunesApp(); //Create instance of Itunes App
-        VideoInfo video = null;
+        public static string Folder = null;
+        public iTunesApp Itunes = new iTunesApp(); //Create instance of Itunes App
+        public VideoInfo Video = null;
+	    public FolderBrowserDialog OutputFolder;
 
         delegate void updatePercentageCallback(double percentage);
 
-        public iPhoneRingtoneMaker()
+        public RingtoneMakerIPhone()
         {
             InitializeComponent();
         }
 
         private void btnOutputDir_Click(object sender, EventArgs e)
         {
-            FolderBrowserDialog outputFolder = new FolderBrowserDialog();
-            outputFolder.RootFolder = Environment.SpecialFolder.DesktopDirectory;
+            OutputFolder = new FolderBrowserDialog
+	            {
+		            RootFolder = Environment.SpecialFolder.DesktopDirectory
+	            };
 
-            if(txtOutputDirectory.Text != "")
+	        if(!string.IsNullOrEmpty(txtOutputDirectory.Text))
             {
-                outputFolder.SelectedPath = txtOutputDirectory.Text;
+                OutputFolder.SelectedPath = txtOutputDirectory.Text;
             }
 
-            if (outputFolder.ShowDialog() == DialogResult.OK)
+            if (OutputFolder.ShowDialog() == DialogResult.OK)
             {
-                txtOutputDirectory.Text = outputFolder.SelectedPath.ToString() + "\\";
-                folder = txtOutputDirectory.Text;
+                txtOutputDirectory.Text = string.Concat(OutputFolder.SelectedPath, "\\");
+                Folder = txtOutputDirectory.Text;
             }
         }
 
@@ -51,70 +49,73 @@ namespace iPhoneRingtoneMaker
 
             createPlaylist();   //Create a temporary playlist in iTunes
 
+	        var outputPath = string.Concat(OutputFolder.SelectedPath, "\\", tbNewFilename.Text);
+
+			var startTime = Convert.ToInt32(txtStartTime.Text);
+			var endTime = Convert.ToInt32(txtEndTime.Text);
+
+			TrimMp3(txtChosenFile.Text, outputPath, TimeSpan.FromSeconds(startTime), TimeSpan.FromSeconds(endTime));
+
             // search for a certain playlist
             IITUserPlaylist playListImSearchingFor = null;
-            foreach (IITPlaylist pl in itunes.LibrarySource.Playlists)
+            foreach (var pl in Itunes.LibrarySource.Playlists.Cast<IITPlaylist>().Where(pl => pl.Name.Equals("Temp")))
             {
-                if (pl.Name.Equals("Temp"))
-                {
-                    playListImSearchingFor = (IITUserPlaylist)pl;
-                    songfile = txtChosenFile.Text;
-                    playListImSearchingFor.AddFile(songfile); // add a song to that playlist
-                }
+	            playListImSearchingFor = (IITUserPlaylist)pl;
+	            songfile = outputPath;
+	            playListImSearchingFor.AddFile(songfile); // add a song to that playlist
             }
+
+			if (playListImSearchingFor == null)
+				throw new Exception("Sorry the playlist was not found correctly");
 
             //Trim selected MP3 by setting the start and end times
             //At the moment it is just taking the first 30 seconds.
             IITTrack track = playListImSearchingFor.Tracks[1];
+	        IITFileOrCDTrack fileTrack = null;
+
+			if (track.Kind == ITTrackKind.ITTrackKindFile)
+			{
+				fileTrack = (IITFileOrCDTrack)track;
+			}
+
+			if(fileTrack == null)
+				throw new FileNotFoundException("There was a problem locating the file you requested.");
+			
             txtTotalLength.Text = Convert.ToString(track.Time);
-            //int length = Convert.ToInt32(txtTotalLength.Text);
-
-            //TODO: Fix Error checking
-           // if (length > 0)
-            //{
-                track.Start = Convert.ToInt32(txtStartTime.Text);
-                track.Finish = Convert.ToInt32(txtEndTime.Text);
-          //  }
-
-          /*  if (!((track.Start < length) && (track.Finish < length)))
-            {
-                track.Start = 0;
-                track.Finish = 30;
-                MessageBox.Show("Invalid times were entered, therefore we will only take the first 30 seconds from the song!", "Invalid Time", MessageBoxButtons.OK, MessageBoxIcon.Error);
-            }*/
-
-            string trackname = null;
-            var result = itunes.ConvertFile2(songfile);
-            trackname = result.trackName;
+			track.Start = Convert.ToInt32(txtStartTime.Text);
+            track.Finish = Convert.ToInt32(txtEndTime.Text);
+         
+			var result = Itunes.ConvertFile2(songfile);
+            
             while (result.InProgress)
             {
-                //Something needs to go here!
+				////Something needs to go here!
             }
-            trackname += ".m4a";
+            var trackname = string.Concat(Path.GetFileNameWithoutExtension(fileTrack.Location), ".m4a");
 
             //Find iTunes Directory
-            string userDirectory = Environment.GetFolderPath(Environment.SpecialFolder.Personal);
-            userDirectory = userDirectory.Replace("Documents", "Music");
-            userDirectory = userDirectory.TrimEnd();
+	        var userDirectory = Path.GetDirectoryName(fileTrack.Location);
 
-            //Find the newly created file
-            string[] files = Directory.GetFiles(userDirectory, "*"+trackname, SearchOption.AllDirectories);
+			if(userDirectory == null || !Directory.Exists(userDirectory))
+				throw new DirectoryNotFoundException(string.Concat("The following directory could not be found", userDirectory));		
+
+			//Find the newly created file
+            var files = Directory.GetFiles(userDirectory, trackname, SearchOption.AllDirectories);
             
             //Delete Playlist
             playListImSearchingFor.Delete();
 
             //Copy File to chosen output Directory
-
-            //If the file already exists, delete it!
+			//If the file already exists, delete it!
             if(File.Exists(txtOutputDirectory.Text + trackname))
             {
                 File.Delete(txtOutputDirectory.Text + trackname);
             }
             
-            File.Move(files[0].ToString(), txtOutputDirectory.Text + trackname);
+            File.Move(files[0], txtOutputDirectory.Text + trackname);
 
-            string trackname2 = trackname.Replace(".m4a", ".m4r");
-            string renamed = txtOutputDirectory.Text + trackname2;
+            var trackname2 = trackname.Replace(".m4a", ".m4r");
+            var renamed = txtOutputDirectory.Text + trackname2;
 
             //Rename File
             //If the file already exists, delete it!
@@ -127,8 +128,8 @@ namespace iPhoneRingtoneMaker
             File.Move(txtOutputDirectory.Text + trackname, renamed);
 
             //Move File
-            string newLocation = files[0].ToString();
-            int last = newLocation.LastIndexOf("\\") + 1;
+            var newLocation = files[0];
+            var last = newLocation.LastIndexOf("\\", StringComparison.Ordinal) + 1;
             newLocation = newLocation.Remove(last);
             newLocation = newLocation + trackname2;
 
@@ -139,7 +140,7 @@ namespace iPhoneRingtoneMaker
             File.Move(renamed, newLocation);
 
             //Close iTunes again
-            itunes.Quit();
+            Itunes.Quit();
 
             //Open File to import into iTunes
             Process.Start(newLocation);            
@@ -147,7 +148,7 @@ namespace iPhoneRingtoneMaker
 
         private void btnChooseFile_Click(object sender, EventArgs e)
         {
-            OpenFileDialog openFileDlg = new OpenFileDialog();
+            var openFileDlg = new OpenFileDialog();
 
             if (openFileDlg.ShowDialog() == DialogResult.OK)
             {
@@ -157,9 +158,8 @@ namespace iPhoneRingtoneMaker
 
         private void txtEndTimeTextChanged(object sender, EventArgs e)
         {
-            int start, end;
-            start = Convert.ToInt32(txtStartTime.Text);
-            end = Convert.ToInt32(txtEndTime.Text);
+	        int start = Convert.ToInt32(txtStartTime.Text);
+            int end = Convert.ToInt32(txtEndTime.Text);
 
             if (start < end)
             {
@@ -181,20 +181,19 @@ namespace iPhoneRingtoneMaker
         public void createPlaylist()
         {
             // create a new playList
-            IITUserPlaylist newPlayList = (IITUserPlaylist)itunes.CreatePlaylist("Temp");
+            var newPlayList = (IITUserPlaylist)Itunes.CreatePlaylist("Temp");
         }
 
         public void createPlaylist(string name)
         {
             // create a new playList
-            IITUserPlaylist newPlayList = (IITUserPlaylist)itunes.CreatePlaylist(name);
+            var newPlayList = (IITUserPlaylist)Itunes.CreatePlaylist(name);
         }
 
         private void txtStartTime_TextChanged(object sender, EventArgs e)
         {
-            int start, end;
-            start = Convert.ToInt32(txtStartTime.Text);
-            end = Convert.ToInt32(txtEndTime.Text);
+	        var start = Convert.ToInt32(txtStartTime.Text);
+            var end = Convert.ToInt32(txtEndTime.Text);
 
             if (start < 0)
             {
@@ -212,11 +211,12 @@ namespace iPhoneRingtoneMaker
 
         private void btnChooseDir_Click(object sender, EventArgs e)
         {
-            FolderBrowserDialog browse = new FolderBrowserDialog();
+            var browse = new FolderBrowserDialog
+	            {
+		            RootFolder = Environment.SpecialFolder.Desktop
+	            };
 
-            browse.RootFolder = Environment.SpecialFolder.Desktop;
-
-            if (browse.ShowDialog() == DialogResult.OK)
+	        if (browse.ShowDialog() == DialogResult.OK)
             {
                 txtMP3Path.Text = browse.SelectedPath.ToString();
             }
@@ -225,33 +225,38 @@ namespace iPhoneRingtoneMaker
         private void btnCreateMP3_Click(object sender, EventArgs e)
         {
             // Our test youtube link
-            string link = txtYouTubeURL.Text;
+            var link = txtYouTubeURL.Text;
 
             /*
              * Get the available video formats.
              * We'll work with them in the video and audio download examples.
              */
-            IEnumerable<VideoInfo> videoInfos = DownloadUrlResolver.GetDownloadUrls(link);
+            var videoInfos = DownloadUrlResolver.GetDownloadUrls(link);
 
             /*
              * We want the first flash (only flash audio extraction is currently supported)
              * video with the highest audio quality.
              * See the VideoFormat enum for more info about the quality.
              */
-            video = videoInfos
+            Video = videoInfos
                 .Where(info => info.CanExtractAudio && info.AudioType == AudioType.Mp3)
                 .OrderByDescending(info => info.AudioBitrate)
                 .First();
 
-            worker.RunWorkerAsync(video);
+            worker.RunWorkerAsync(Video);
 
         }
 
         private void worker_DoWork(object sender, DoWorkEventArgs e)
         {
             //Set the path to a temp directory on the desktop.
-            string folderpath = Environment.GetFolderPath(Environment.SpecialFolder.Desktop);
+            var folderpath = Environment.GetFolderPath(Environment.SpecialFolder.Desktop);
             folderpath += "\\temp\\";
+
+			if (!Directory.Exists(folderpath))
+			{
+				Directory.CreateDirectory(folderpath);
+			}
             /*
               * Create the audio downloader.
               * The first argument is the video where the audio should be extracted from.
@@ -259,9 +264,9 @@ namespace iPhoneRingtoneMaker
               * Automatic video title infering will be supported later.
               */
             //var audioDownloader = new AudioDownloader(video, folderpath + video.AudioExtension);
-            var audioDownloader = new AudioDownloader(video, Path.Combine(folderpath, video.Title + video.AudioExtension));
+            var audioDownloader = new AudioDownloader(Video, Path.Combine(folderpath, Video.Title + Video.AudioExtension));
             // Register the ProgressChanged event and print the current progress
-            audioDownloader.ProgressChanged += (senders, args) => updatePercentage(args.ProgressPercentage);
+            audioDownloader.DownloadProgressChanged += (senders, args) => updatePercentage(args.ProgressPercentage);
             
             /*
              * Execute the audio downloader.
@@ -292,5 +297,24 @@ namespace iPhoneRingtoneMaker
             //Display a dialog when the extraction is complete.
             MessageBox.Show("MP3 file extracted!", "Complete", MessageBoxButtons.OK, MessageBoxIcon.Information);
         }
+
+		void TrimMp3(string inputPath, string outputPath, TimeSpan? begin, TimeSpan? end)
+{
+    if (begin.HasValue && end.HasValue && begin > end)
+        throw new ArgumentOutOfRangeException("end", "end should be greater than begin");
+
+    using (var reader = new Mp3FileReader(inputPath))
+    using (var writer = File.Create(outputPath))
+    {           
+        Mp3Frame frame;
+        while ((frame = reader.ReadNextFrame()) != null)
+        if (reader.CurrentTime >= begin || !begin.HasValue)
+        {
+            if (reader.CurrentTime <= end || !end.HasValue)
+                writer.Write(frame.RawData,0,frame.RawData.Length);         
+            else break;
+        }
+    }
+}
     }
 }
